refactor each component into own file
