+++ ecommerce/frontend/src/components/CategoryBar.js
@@ -1,0 +1,17 @@
+import { Menu } from 'antd'
+import { Link } from 'react-router-dom'
+
+export const CategoryBar = ({ categories = [] }) => {
+    return (
+        <Menu mode="horizontal" className="category-menu">
+            {categories.map(({ key, title }) => {
+                const url = `/categories/${key}`
+                return (
+                    <Menu.Item className="category-menu__menu-item" key={key}>
+                        <Link to={url}>{title}</Link>
+                    </Menu.Item>
+                )
+            })}
+        </Menu>
+    )
+}
