+++ ecommerce/frontend/src/components/LoginPage.js
@@ -1,0 +1,3 @@
+export const LoginPage = () => {
+    return 'login'
+}
