+++ ecommerce/frontend/src/components/SignupPage.js
@@ -1,0 +1,3 @@
+export const SignupPage = () => {
+    return 'signup'
+}
