+++ ecommerce/frontend/src/components/layout/MainLayout.js
@@ -1,0 +1,14 @@
+import { Layout } from 'antd'
+import { Header } from './Header'
+import { Content } from './Content'
+import { Footer } from './Footer'
+
+export const MainLayout = () => {
+    return (
+        <Layout>
+            <Header />
+            <Content />
+            <Footer />
+        </Layout>
+    )
+}
